refactoring

+++ eve-tool-spray/src/main/scala/info/golushkov/eve/tool/akka/models/Order.scala
@@ -7,7 +7,7 @@ case class Order(
                   lastUpdate: LocalDateTime,
                   isBuy: Boolean,
                   locationId: Long,
-                  regionId: Int,
+                  regionId: Long,
                   price: Double,
                   itemId: Int,
                   remain: Int,
